removing code smell
