+++ app/client/views/settings/users/edit/schema.js
@@ -1,4 +1,5 @@
 import SimpleSchema from 'simpl-schema';
+import { deactivateOnSchema } from '../commonSchemas';
 
 EditUserSchema = new SimpleSchema({
   email: {
@@ -10,23 +11,5 @@ EditUserSchema = new SimpleSchema({
     type: Boolean,
     defaultValue: false,
   },
-  deactivateOn: {
-    type: Date,
-    optional: true,
-    min: function() {
-      const date = new Date();
-      const utcDate = Date.UTC(
-        date.getFullYear(),
-        date.getMonth(),
-        date.getDate(),
-        0,
-        0,
-        0,
-        0
-      );
-
-      return new Date(utcDate);
-    },
-    max: new Date('2050'),
-  },
+  deactivateOn: deactivateOnSchema,
 });
