+++ app/client/views/settings/users/new/schema.js
@@ -1,4 +1,5 @@
 import SimpleSchema from 'simpl-schema';
+import { deactivateOnSchema } from '../commonSchemas';
 
 NewUserSchema = new SimpleSchema({
   "email": {
@@ -16,23 +17,5 @@ NewUserSchema = new SimpleSchema({
     label: "System administrator"
   },
 
-  "deactivateOn": {
-    type: Date,
-    optional: true,
-    min: function() {
-      const date = new Date();
-      const utcDate = Date.UTC(
-        date.getFullYear(),
-        date.getMonth(),
-        date.getDate(),
-        0,
-        0,
-        0,
-        0
-      );
-
-      return new Date(utcDate);
-    },
-    max: new Date('2050'),
-  },
+  "deactivateOn": deactivateOnSchema,
 });
